+++ common/src/main/java/com/jaquadro/minecraft/storagedrawers/config/StorageBlacklist.java
@@ -34,11 +34,13 @@ public class StorageBlacklist extends ConfigItemList
 
     @Override
     protected void logRegisterNamespace (@NotNull String namespace) {
-        ModServices.log.info("New blacklisted storage namespace " + namespace);
+        if (ModCommonConfig.INSTANCE.GENERAL.logStartupActivity.get())
+            ModServices.log.info("New denied storage namespace " + namespace);
     }
 
     @Override
     protected void logRegisterItem (@NotNull ItemStack item) {
-        ModServices.log.info("New blacklisted storage item " + item.getItem());
+        if (ModCommonConfig.INSTANCE.GENERAL.logStartupActivity.get())
+            ModServices.log.info("New denied storage item " + item.getItem());
     }
 }