+++ fabric/src/main/java/com/jaquadro/minecraft/storagedrawers/StorageDrawers.java
@@ -40,5 +40,6 @@ public class StorageDrawers implements ModInitializer
         CompTierRegistry.INSTANCE.initialize();
         StorageBlacklist.INSTANCE.initialize();
         MaterialBlacklist.INSTANCE.initialize();
+        ConversionRegistry.INSTANCE.initialize();
     }
 }