+++ forge/src/main/java/com/jaquadro/minecraft/storagedrawers/StorageDrawers.java
@@ -78,6 +78,7 @@ public class StorageDrawers
         CompTierRegistry.INSTANCE.initialize();
         StorageBlacklist.INSTANCE.initialize();
         MaterialBlacklist.INSTANCE.initialize();
+        ConversionRegistry.INSTANCE.initialize();
         PlatformCapabilities.initHandlers();
 
         //oreDictRegistry = new OreDictRegistry();