+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 [12.12.1]
 - Fixed meta blockstate loading errors (though they were harmless)
+- Fixed conversion rules not loading
+- Added option to disable logging startup rules, deny lists, etc.
 
 [12.12.0]
 - New config file (storagedrawers-common-v2.toml)